admin: refacto add hasAccessToCertifAction helper
