+++ admin/app/controllers/authenticated/sessions/list/to-be-published.js
@@ -7,16 +7,8 @@ import get from 'lodash/get';
 export default class SessionToBePublishedController extends Controller {
   @service notifications;
   @service store;
-  @service currentUser;
+  @service accessControl;
   @tracked shouldShowModal = false;
-
-  get isCurrentUserAllowedToPublishSession() {
-    return (
-      this.currentUser.adminMember.isSuperAdmin ||
-      this.currentUser.adminMember.isSupport ||
-      this.currentUser.adminMember.isCertif
-    );
-  }
 
   @action
   async publishSession(toBePublishedSession) {
