+++ admin/app/controllers/authenticated/sessions/session/certifications.js
@@ -9,7 +9,7 @@ import { action, computed } from '@ember/object';
 export default class ListController extends Controller {
   @service notifications;
   @service store;
-  @service currentUser;
+  @service accessControl;
 
   @tracked displayConfirm = false;
   @tracked confirmMessage = null;
@@ -18,14 +18,6 @@ export default class ListController extends Controller {
   get canPublish() {
     return !some(this.model.juryCertificationSummaries.toArray(), (certif) =>
       ['error', 'started'].includes(certif.status)
-    );
-  }
-
-  get isCurrentUserAllowedToPublishSession() {
-    return (
-      this.currentUser.adminMember.isSuperAdmin ||
-      this.currentUser.adminMember.isSupport ||
-      this.currentUser.adminMember.isCertif
     );
   }
