+++ admin/app/templates/authenticated/sessions/list/to-be-published.hbs
@@ -1,6 +1,6 @@
 {{page-title "Sessions à publier"}}
 {{#if @model.length}}
-  {{#if this.isCurrentUserAllowedToPublishSession}}
+  {{#if this.accessControl.hasAccessToCertificationActionsScope}}
     <div class="session-to-be-publish__publish-all">
       <PixButton @triggerAction={{this.showConfirmModal}}>Publier toutes les sessions</PixButton>
     </div>
