+++ admin/tests/integration/components/routes/authenticated/to-be-published-sessions/list-items_test.js
@@ -3,14 +3,18 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render, clickByName } from '@1024pix/ember-testing-library';
 import hbs from 'htmlbars-inline-precompile';
 import sinon from 'sinon';
+import Service from '@ember/service';
 
 module('Integration | Component | routes/authenticated/to-be-published-sessions | list-items', function (hooks) {
   setupRenderingTest(hooks);
 
   test('it should display to be published sessions list', async function (assert) {
     // given
-    const currentUser = this.owner.lookup('service:currentUser');
-    currentUser.adminMember = { isCertif: true };
+    class SessionStub extends Service {
+      hasAccessToCertificationActionsScope = true;
+    }
+    this.owner.register('service:accessControl', SessionStub);
+
     const firstSession = {
       id: '1',
       certificationCenterName: 'Centre SCO des Anne-Étoiles',
@@ -48,8 +52,11 @@ module('Integration | Component | routes/authenticated/to-be-published-sessions 
 
   test('it should "Aucun résultat" if there are no sessions to show', async function (assert) {
     // given
-    const currentUser = this.owner.lookup('service:currentUser');
-    currentUser.adminMember = { isCertif: true };
+    class SessionStub extends Service {
+      hasAccessToCertificationActionsScope = true;
+    }
+    this.owner.register('service:accessControl', SessionStub);
+
     this.toBePublishedSessions = [];
 
     // when
@@ -62,10 +69,13 @@ module('Integration | Component | routes/authenticated/to-be-published-sessions 
   });
 
   module('Publish a session', function () {
-    test('it should not show button if current user is metier', async function (assert) {
+    test('it should not show button if current user does not have the right', async function (assert) {
       // given
-      const currentUser = this.owner.lookup('service:currentUser');
-      currentUser.adminMember = { isMetier: true };
+      class SessionStub extends Service {
+        hasAccessToCertificationActionsScope = false;
+      }
+      this.owner.register('service:accessControl', SessionStub);
+
       const session = {
         id: '1',
         certificationCenterName: 'Centre SCO des Anne-Étoiles',
@@ -86,8 +96,11 @@ module('Integration | Component | routes/authenticated/to-be-published-sessions 
 
     test('it should show confirmation modal when one clicks on "Publier" button', async function (assert) {
       // given
-      const currentUser = this.owner.lookup('service:currentUser');
-      currentUser.adminMember = { isCertif: true };
+
+      class SessionStub extends Service {
+        hasAccessToCertificationActionsScope = true;
+      }
+      this.owner.register('service:accessControl', SessionStub);
 
       const session = {
         id: '1',
@@ -110,8 +123,10 @@ module('Integration | Component | routes/authenticated/to-be-published-sessions 
 
     test('it should call provided publishModel "action" with the right published session as argument', async function (assert) {
       // given
-      const currentUser = this.owner.lookup('service:currentUser');
-      currentUser.adminMember = { isCertif: true };
+      class SessionStub extends Service {
+        hasAccessToCertificationActionsScope = true;
+      }
+      this.owner.register('service:accessControl', SessionStub);
 
       const session = {
         id: '1',
