+++ admin/tests/unit/services/access-control-test.js
@@ -46,7 +46,7 @@ module('Unit | Service | access-control', function (hooks) {
   });
 
   module('#hasAccessToOrganizationActionsScope', function () {
-    test('should be true if user is Super Admin', function (assert) {
+    test('should be true if current admin member is Super Admin', function (assert) {
       // given
       const currentUser = this.owner.lookup('service:currentUser');
       currentUser.adminMember = { isSuperAdmin: true };
@@ -88,6 +88,52 @@ module('Unit | Service | access-control', function (hooks) {
 
       // when / then
       assert.false(service.hasAccessToOrganizationActionsScope);
+    });
+  });
+
+  module('#hasAccessToCertificationActionsScope', function () {
+    test('should be true if current admin member is Super admin', function (assert) {
+      // given
+      const currentUser = this.owner.lookup('service:currentUser');
+      currentUser.adminMember = { isSuperAdmin: true };
+
+      const service = this.owner.lookup('service:access-control');
+
+      // when / then
+      assert.true(service.hasAccessToCertificationActionsScope);
+    });
+
+    test('should be true if user is Support', function (assert) {
+      // given
+      const currentUser = this.owner.lookup('service:currentUser');
+      currentUser.adminMember = { isSupport: true };
+
+      const service = this.owner.lookup('service:access-control');
+
+      // when / then
+      assert.true(service.hasAccessToCertificationActionsScope);
+    });
+
+    test('should be true if user is Certif', function (assert) {
+      // given
+      const currentUser = this.owner.lookup('service:currentUser');
+      currentUser.adminMember = { isCertif: true };
+
+      const service = this.owner.lookup('service:access-control');
+
+      // when / then
+      assert.true(service.hasAccessToCertificationActionsScope);
+    });
+
+    test('should be false if user is Metier', function (assert) {
+      // given
+      const currentUser = this.owner.lookup('service:currentUser');
+      currentUser.adminMember = { isMetier: true };
+
+      const service = this.owner.lookup('service:access-control');
+
+      // when / then
+      assert.false(service.hasAccessToCertificationActionsScope);
     });
   });
 });
